refactor : 컴포넌트 추상화를 위한 컴포넌트 분리
