+++ src/components/CurrentVote/CurrentVoteInfo.tsx
@@ -3,7 +3,7 @@ import * as timeago from 'timeago.js';
 import ko from 'timeago.js/lib/lang/ko';
 
 import { pictureCurrentIdx, pictureResultState, voteResultState } from '../../recoil/maker/atom';
-import CurrentVoteInfoLayout from './Layout/CurrentVoteInfoLayout';
+import { CurrentVoteInfoLayout } from './Layout';
 
 const CurrentVoteInfo = () => {
   const voteInfoData = useRecoilValue(voteResultState);
