+++ src/components/CurrentVote/OriginVote/index.ts
@@ -1,0 +1,1 @@
+export { default as OriginVotePicture } from './OriginVotePicture';
